+++ ForYouAndMe/Classes/UI/Navigator/ReflectionSectionCoordinator.swift
@@ -216,7 +216,7 @@ extension ReflectionSectionCoordinator: ReflectionViewCoordinator {
         case .text:
             self.sendResult(taskResult: taskResultData, presenter: presenter)
             return
-        case .eaten, .doses, .weNoticed:
+        case .eaten, .doses, .weNoticed, .hotFlash:
             return
         }
     }