+++ ForYouAndMe/Classes/UI/ViewControllers/EmojiPopUpViewController.swift
@@ -18,6 +18,7 @@ enum EmojiTagCategory: String, CaseIterable {
     case iHaveEaten = "i_have_eaten"
     case iHaveNoticed = "i_have_noticed"
     case weHaveNoticed = "we_have_noticed"
+    case hotFlash = "hot_flash"
     case none
 }
 
@@ -69,7 +70,7 @@ final class EmojiPopupViewController: UIViewController {
         view.backgroundColor = UIColor.black.withAlphaComponent(0.6)
         
         let cardView = UIView()
-        cardView.backgroundColor = ColorPalette.color(withType: .gradientPrimaryEnd)
+        cardView.backgroundColor = ColorPalette.color(withType: .secondary)
         cardView.layer.cornerRadius = 20
         cardView.clipsToBounds = true
         view.addSubview(cardView)